Decimate*: Refine face count adjustment

+++ src/core/LoopSubface.hpp
@@ -8,7 +8,7 @@ namespace subface {
 
 class LoopSubface {
     int level_ = 0;
-    int result_face_count_ = 0;
+    size_t result_face_count_ = 0;
 
     std::vector<glm::vec3> origin_positions_;
     std::vector<uint32_t> origin_indexes_;
